Move ssr audiusSdk to utils file
